+++ src/main/webapp/wise5/authoringTool/node/advanced/node-advanced-authoring.component.ts
@@ -1,5 +1,4 @@
 import { ConfigService } from "../../../services/configService";
-import { NotificationService } from "../../../services/notificationService";
 import { TeacherDataService } from "../../../services/teacherDataService";
 import { TeacherProjectService } from "../../../services/teacherProjectService";
 import { UtilService } from "../../../services/utilService";
@@ -8,7 +7,6 @@ import { TagService } from "../../../services/tagService";
 class NodeAdvancedAuthoringController {
 
   $translate: any;
-  constraintActions: any[];
   items: any[];
   node: any;
   nodeId: string;
@@ -19,19 +17,15 @@ class NodeAdvancedAuthoringController {
   createBranchNodeId: string;
   createBranchNumberOfBranches: any;
   createBranchCriterion: any;
-  removalConditionals: any[];
-  removalCriteria: any;
-  showConstraints: boolean = false;
   showCreateBranch: boolean = false;
   showEditTransitions: boolean = false;
   transitionCriterias: any;
 
-  static $injector = ['$filter', '$state', '$timeout', 'ConfigService', 'NotificationService',
+  static $injector = ['$filter', '$state', '$timeout', 'ConfigService',
       'ProjectService', 'TagService', 'TeacherDataService', 'UtilService'];
 
   constructor(private $filter: any, private $state: any, private $timeout: any,
       private ConfigService: ConfigService,
-      private NotificationService: NotificationService,
       private ProjectService: TeacherProjectService,
       private TagService: TagService,
       private TeacherDataService: TeacherDataService,
@@ -57,241 +51,6 @@ class NodeAdvancedAuthoringController {
       {
         value: 'tag',
         text: this.$translate('tag')
-      }
-    ];
-    this.constraintActions = [
-      {
-        value: '',
-        text: this.$translate('pleaseChooseAnAction')
-      },
-      {
-        value: 'makeAllNodesAfterThisNotVisitable',
-        text: this.$translate('makeAllNodesAfterThisNotVisitable')
-      },
-      {
-        value: 'makeAllNodesAfterThisNotVisible',
-        text: this.$translate('makeAllNodesAfterThisNotVisible')
-      },
-      {
-        value: 'makeAllOtherNodesNotVisitable',
-        text: this.$translate('makeAllOtherNodesNotVisitable')
-      },
-      {
-        value: 'makeAllOtherNodesNotVisible',
-        text: this.$translate('makeAllOtherNodesNotVisible')
-      },
-      {
-        value: 'makeThisNodeNotVisitable',
-        text: this.$translate('makeThisNodeNotVisitable')
-      },
-      {
-        value: 'makeThisNodeNotVisible',
-        text: this.$translate('makeThisNodeNotVisible')
-      }
-    ];
-    this.removalConditionals = [
-      {
-        value: 'all',
-        text: this.$translate('all')
-      },
-      {
-        value: 'any',
-        text: this.$translate('any')
-      }
-    ];
-    this.removalCriteria = [
-      {
-        value: '',
-        text: this.$translate('pleaseChooseARemovalCriteria')
-      },
-      {
-        value: 'isCompleted',
-        text: this.$translate('isCompleted'),
-        params: [
-          {
-            value: 'nodeId',
-            text: this.$translate('step')
-          }
-        ]
-      },
-      {
-        value: 'score',
-        text: this.$translate('SCORE'),
-        params: [
-          {
-            value: 'nodeId',
-            text: this.$translate('step')
-          },
-          {
-            value: 'component',
-            text: this.$translate('component')
-          },
-          {
-            value: 'scores',
-            text: this.$translate('scoresParens')
-          }
-        ]
-      },
-      {
-        value: 'branchPathTaken',
-        text: this.$translate('branchPathTaken'),
-        params: [
-          {
-            value: 'fromNodeId',
-            text: this.$translate('fromStep')
-          },
-          {
-            value: 'toNodeId',
-            text: this.$translate('toStep')
-          }
-        ]
-      },
-      {
-        value: 'choiceChosen',
-        text: this.$translate('choiceChosen'),
-        params: [
-          {
-            value: 'nodeId',
-            text: this.$translate('step')
-          },
-          {
-            value: 'componentId',
-            text: this.$translate('component')
-          },
-          {
-            value: 'choiceIds',
-            text: this.$translate('choices')
-          }
-        ]
-      },
-      {
-        value: 'isCorrect',
-        text: this.$translate('IS_CORRECT'),
-        params: [
-          {
-            value: 'nodeId',
-            text: this.$translate('step')
-          },
-          {
-            value: 'componentId',
-            text: this.$translate('component')
-          }
-        ]
-      },
-      {
-        value: 'usedXSubmits',
-        text: this.$translate('usedXSubmits'),
-        params: [
-          {
-            value: 'nodeId',
-            text: this.$translate('step')
-          },
-          {
-            value: 'componentId',
-            text: this.$translate('component')
-          },
-          {
-            value: 'requiredSubmitCount',
-            text: this.$translate('requiredSubmitCount')
-          }
-        ]
-      },
-      {
-        value: 'isVisible',
-        text: this.$translate('isVisible'),
-        params: [
-          {
-            value: 'nodeId',
-            text: this.$translate('step')
-          }
-        ]
-      },
-      {
-        value: 'isVisitable',
-        text: this.$translate('isVisitable'),
-        params: [
-          {
-            value: 'nodeId',
-            text: this.$translate('step')
-          }
-        ]
-      },
-      {
-        value: 'isVisited',
-        text: this.$translate('isVisited'),
-        params: [
-          {
-            value: 'nodeId',
-            text: this.$translate('step')
-          }
-        ]
-      },
-      {
-        value: 'wroteXNumberOfWords',
-        text: this.$translate('wroteXNumberOfWords'),
-        params: [
-          {
-            value: 'nodeId',
-            text: this.$translate('step')
-          },
-          {
-            value: 'componentId',
-            text: this.$translate('component')
-          },
-          {
-            value: 'requiredNumberOfWords',
-            text: this.$translate('requiredNumberOfWords')
-          }
-        ]
-      },
-      {
-        value: 'addXNumberOfNotesOnThisStep',
-        text: this.$translate('addXNumberOfNotesOnThisStep'),
-        params: [
-          {
-            value: 'nodeId',
-            text: this.$translate('step')
-          },
-          {
-            value: 'requiredNumberOfNotes',
-            text: this.$translate('requiredNumberOfNotes')
-          }
-        ]
-      },
-      {
-        value: 'fillXNumberOfRows',
-        text: this.$translate('fillXNumberOfRows'),
-        params: [
-          {
-            value: 'nodeId',
-            text: this.$translate('step')
-          },
-          {
-            value: 'componentId',
-            text: this.$translate('component')
-          },
-          {
-            value: 'requiredNumberOfFilledRows',
-            defaultValue: null,
-            text: this.$translate('requiredNumberOfFilledRowsNotIncludingHeaderRow')
-          },
-          {
-            value: 'tableHasHeaderRow',
-            defaultValue: true,
-            text: this.$translate('tableHasHeaderRow')
-          },
-          {
-            value: 'requireAllCellsInARowToBeFilled',
-            defaultValue: true,
-            text: this.$translate('requireAllCellsInARowToBeFilled')
-          }
-        ]
-      },
-      {
-        value: 'teacherRemoval',
-        text: this.$translate('teacherRemoval'),
-        params: [
-        ]
       }
     ];
     this.transitionCriterias = [
@@ -355,12 +114,6 @@ class NodeAdvancedAuthoringController {
         this.showEditTransitionsView();
       });
     }
-
-    if (this.$state.current.name === 'root.at.project.nodeConstraints') {
-      this.$timeout(() => {
-        this.showEditConstraintsView();
-      });
-    }
   }
 
   goBack() {
@@ -379,8 +132,7 @@ class NodeAdvancedAuthoringController {
   }
 
   showEditConstraintsView() {
-    this.hideAllViews();
-    this.showConstraints = true;
+    this.$state.go('root.at.project.node.advanced.constraint');
   }
 
   showGeneralAdvancedView() {
@@ -393,7 +145,6 @@ class NodeAdvancedAuthoringController {
 
   hideAllViews() {
     this.showEditTransitions = false;
-    this.showConstraints = false;
     this.showCreateBranch = false;
   }
 
@@ -464,25 +215,6 @@ class NodeAdvancedAuthoringController {
       const transitionCriterias = transition.criteria;
       if (transitionCriterias != null) {
         transitionCriterias.splice(transitionCriteriaIndex, 1);
-      }
-      this.ProjectService.saveProject();
-    }
-  }
-
-  addRemovalCriteria(constraint) {
-    const removalCriteria = {
-      name: '',
-      params: {}
-    };
-    constraint.removalCriteria.push(removalCriteria);
-    this.ProjectService.saveProject();
-  }
-
-  deleteRemovalCriteria(constraint, removalCriteriaIndex) {
-    if (confirm(this.$translate('areYouSureYouWantToDeleteThisRemovalCriteria'))) {
-      const removalCriteria = constraint.removalCriteria;
-      if (removalCriteria != null) {
-        removalCriteria.splice(removalCriteriaIndex, 1);
       }
       this.ProjectService.saveProject();
     }
@@ -578,105 +310,6 @@ class NodeAdvancedAuthoringController {
       this.ProjectService.deleteTransition(this.node, transition);
       this.authoringViewNodeChanged();
     }
-  }
-
-  addConstraintAndScrollToBottom() {
-    const newNodeConstraintId = this.addConstraint();
-    this.$timeout(() => {
-      this.ProjectService.scrollToBottomOfPage();
-      this.UtilService.temporarilyHighlightElement(newNodeConstraintId);
-    });
-  }
-
-  addConstraint() {
-    const newNodeConstraintId = this.getNewNodeConstraintId(this.nodeId);
-    const constraint = {
-      id: newNodeConstraintId,
-      action: '',
-      targetId: this.nodeId,
-      removalConditional: 'any',
-      removalCriteria: [{
-        name: '',
-        params: {}
-      }]
-    };
-    if (this.node.constraints == null) {
-      this.node.constraints = [];
-    }
-    this.node.constraints.push(constraint);
-    this.ProjectService.saveProject();
-    return newNodeConstraintId;
-  }
-
-  getNewNodeConstraintId(nodeId) {
-    let newNodeConstraintId = null;
-    const usedConstraintIds = [];
-    const node = this.ProjectService.getNodeById(nodeId);
-    if (node != null && node.constraints != null) {
-      for (const constraint of node.constraints) {
-        if (constraint != null) {
-          usedConstraintIds.push(constraint.id);
-        }
-      }
-    }
-    let constraintCounter = 1;
-    while (newNodeConstraintId == null) {
-      let potentialNewNodeConstraintId = nodeId + 'Constraint' + constraintCounter;
-      if (usedConstraintIds.indexOf(potentialNewNodeConstraintId) == -1) {
-        newNodeConstraintId = potentialNewNodeConstraintId;
-      } else {
-        constraintCounter++;
-      }
-    }
-    return newNodeConstraintId;
-  }
-
-  deleteConstraint(constraintIndex) {
-    if (confirm(this.$translate('areYouSureYouWantToDeleteThisConstraint'))) {
-      const node = this.ProjectService.getNodeById(this.nodeId);
-      const constraints = node.constraints;
-      if (constraints != null) {
-        constraints.splice(constraintIndex, 1);
-      }
-      this.ProjectService.saveProject();
-    }
-  }
-
-  constraintRemovalCriteriaNodeIdChanged(criteria) {
-    criteria.params.componentId = '';
-    this.authoringViewNodeChanged();
-  }
-
-  constraintRemovalCriteriaComponentIdChanged(criteria) {
-    this.authoringViewNodeChanged();
-  }
-
-  removalCriteriaNameChanged(criteria) {
-    criteria.params = {};
-    const params = this.getRemovalCriteriaParamsByName(criteria.name);
-    if (params != null) {
-      for (const paramObject of params) {
-        const value = paramObject.value;
-        if (paramObject.hasOwnProperty('defaultValue')) {
-          criteria.params[value] = paramObject.defaultValue;
-        } else {
-          criteria.params[value] = '';
-        }
-        if (value === 'nodeId') {
-          criteria.params[value] = this.nodeId;
-        }
-      }
-    }
-    this.authoringViewNodeChanged();
-  }
-
-  getRemovalCriteriaParamsByName(name) {
-    for (const singleRemovalCriteria of this.removalCriteria) {
-      if (singleRemovalCriteria.value === name) {
-        return singleRemovalCriteria.params;
-      }
-    }
-    return [];
   }
 
   populateBranchAuthoring() {
