+++ src/main/webapp/wise5/teacher/teacher-angular-js-module.ts
@@ -21,6 +21,7 @@ import ExportVisitsController from '../classroomMonitor/dataExport/exportVisitsC
 import MilestonesAuthoringController from '../authoringTool/milestones/milestonesAuthoringController';
 import MilestonesController from '../classroomMonitor/milestones/milestonesController';
 import { NodeAdvancedAuthoringComponent } from '../authoringTool/node/advanced/node-advanced-authoring.component';
+import { NodeAdvancedConstraintAuthoringComponent } from '../authoringTool/node/advanced/constraint/node-advanced-constraint-authoring.component';
 import { NodeAdvancedGeneralAuthoringComponent } from '../authoringTool/node/advanced/general/node-advanced-general-authoring.component';
 import { NodeAdvancedJsonAuthoringComponent } from '../authoringTool/node/advanced/json/node-advanced-json-authoring.component';
 import NodeAuthoringController from '../authoringTool/node/nodeAuthoringController';
@@ -98,6 +99,7 @@ export function createTeacherAngularJSModule() {
     .service('TeacherDataService', downgradeInjectable(TeacherDataService))
     .service('TeacherWebSocketService', downgradeInjectable(TeacherWebSocketService))
     .component('nodeAdvancedAuthoringComponent', NodeAdvancedAuthoringComponent)
+    .component('nodeAdvancedConstraintAuthoringComponent', NodeAdvancedConstraintAuthoringComponent)
     .component('nodeAdvancedGeneralAuthoringComponent', NodeAdvancedGeneralAuthoringComponent)
     .component('nodeAdvancedJsonAuthoringComponent', NodeAdvancedJsonAuthoringComponent)
     .directive('advancedProjectAuthoringComponent', downgradeComponent(
@@ -227,9 +229,9 @@ export function createTeacherAngularJSModule() {
           url: '/json',
           component: 'nodeAdvancedJsonAuthoringComponent'
         })
-        .state('root.at.project.nodeConstraints', {
-          url: '/advanced',
-          component: 'nodeAdvancedAuthoringComponent'
+        .state('root.at.project.node.advanced.constraint', {
+          url: '/constraint',
+          component: 'nodeAdvancedConstraintAuthoringComponent'
         })
         .state('root.at.project.nodeEditPaths', {
           url: '/advanced',
